First (major) refactoring.
All hardcoded data is now served from a server, and the controllers
and directives at main.js were moved to their corresponding files.

That might change though, if the 'demo' becomes too big to justify
a more modular approach, like, a folder for each functionality instead
of three mega files for the whole app (one for services, one for
controllers, and the other forr directives).